+++ e2e/commands/build.e2e.js
@@ -32,4 +32,85 @@ describe('bit build', function () {
       expect(buildOutput).to.have.string(path.normalize('local/dist/bar/foo.js'));
     });
   });
+  describe('with dist attribute populated', () => {
+    let localConsumerFiles;
+    let clonedScope;
+    before(() => {
+      helper.setNewLocalAndRemoteScopes();
+      helper.importCompiler();
+      helper.createComponent(path.normalize('src/bar'), 'foo.js');
+      helper.addComponent(path.normalize('src/bar/foo.js'));
+      clonedScope = helper.cloneLocalScope();
+    });
+    describe('as author', () => {
+      describe('with dist.entry populated', () => {
+        before(() => {
+          helper.modifyFieldInBitJson('dist', { entry: 'src' });
+          helper.build();
+          localConsumerFiles = helper.getConsumerFiles('*.{js,ts}', false);
+        });
+        it('should write the dists files without the dist.entry part', () => {
+          expect(localConsumerFiles).to.include(path.join('dist', 'bar', 'foo.js'));
+          expect(localConsumerFiles).to.not.include(path.join('dist', 'src', 'bar', 'foo.js'));
+        });
+      });
+      describe('with dist.entry and dist.target populated', () => {
+        before(() => {
+          helper.getClonedLocalScope(clonedScope);
+          helper.modifyFieldInBitJson('dist', { entry: 'src', target: 'my-dist' });
+          helper.build();
+          localConsumerFiles = helper.getConsumerFiles('*.{js,ts}', false);
+        });
+        it('should write the dists files inside dist.target dir and without the dist.entry part', () => {
+          expect(localConsumerFiles).to.include(path.join('my-dist', 'bar', 'foo.js'));
+          expect(localConsumerFiles).to.not.include(path.join('my-dist', 'src', 'bar', 'foo.js'));
+          expect(localConsumerFiles).to.not.include(path.join('dist', 'src', 'bar', 'foo.js'));
+          expect(localConsumerFiles).to.not.include(path.join('dist', 'bar', 'foo.js'));
+        });
+      });
+    });
+    describe('as imported', () => {
+      before(() => {
+        helper.getClonedLocalScope(clonedScope);
+        helper.commitAllComponents();
+        helper.exportAllComponents();
+        helper.reInitLocalScope();
+        helper.addRemoteScope();
+        helper.importComponent('bar/foo');
+        clonedScope = helper.cloneLocalScope();
+      });
+      describe('with dist.entry populated', () => {
+        before(() => {
+          helper.modifyFieldInBitJson('dist', { entry: 'src' });
+          helper.build('bar/foo');
+          localConsumerFiles = helper.getConsumerFiles('*.{js,ts}', false);
+        });
+        it('should write the dists files without the dist.entry part and without the originallySharedDirectory part', () => {
+          expect(localConsumerFiles).to.include(path.join('dist', 'components', 'bar', 'foo', 'foo.js'));
+          expect(localConsumerFiles).to.not.include(path.join('dist', 'components', 'bar', 'foo', 'src', 'foo.js')); // dist.entry
+          expect(localConsumerFiles).to.not.include(path.join('dist', 'components', 'bar', 'foo', 'bar', 'foo.js')); // originallyShared
+          expect(localConsumerFiles).to.not.include(
+            path.join('dist', 'components', 'bar', 'foo', 'bar', 'src', 'foo.js')
+          ); // both
+        });
+      });
+      describe('with dist.entry and dist.target populated', () => {
+        before(() => {
+          helper.getClonedLocalScope(clonedScope);
+          helper.modifyFieldInBitJson('dist', { entry: 'src', target: 'my-dist' });
+          helper.build('bar/foo');
+          localConsumerFiles = helper.getConsumerFiles('*.{js,ts}', false);
+        });
+        it('should write the dists files inside dist.target dir and without the dist.entry part', () => {
+          expect(localConsumerFiles).to.include(path.join('my-dist', 'components', 'bar', 'foo', 'foo.js'));
+          expect(localConsumerFiles).to.not.include(path.join('dist', 'components', 'bar', 'foo', 'foo.js')); // default dist.target
+          expect(localConsumerFiles).to.not.include(path.join('my-dist', 'components', 'bar', 'foo', 'src', 'foo.js')); // dist.entry
+          expect(localConsumerFiles).to.not.include(path.join('my-dist', 'components', 'bar', 'foo', 'bar', 'foo.js')); // originallyShared
+          expect(localConsumerFiles).to.not.include(
+            path.join('my-dist', 'components', 'bar', 'foo', 'bar', 'src', 'foo.js')
+          ); // both
+        });
+      });
+    });
+  });
 });